requirements.txt
discord.py
python-dotenv
azure-core
azure-data-tables
fuzzywuzzy